+++ src/tui/util.rs
@@ -1,0 +1,164 @@
+//! Utilities for signal and message handling for the TUI. Most non-core
+//! functionality is spun out into this module.
+
+use crate::{
+    template::Prompt,
+    tui::{
+        message::{Message, MessageSender},
+        view::Confirm,
+    },
+    util::ResultExt,
+};
+use anyhow::Context;
+use futures::{future, FutureExt};
+use std::io;
+use tokio::{fs::OpenOptions, io::AsyncWriteExt, sync::oneshot};
+use tracing::{debug, info, warn};
+
+/// Listen for any exit signals, and return `Ok(())` when any signal is
+/// received. This can only fail during initialization.
+#[cfg(unix)]
+pub async fn signals() -> anyhow::Result<()> {
+    use itertools::Itertools;
+    use tokio::signal::unix::{signal, Signal, SignalKind};
+
+    let signals: Vec<(Signal, SignalKind)> = [
+        SignalKind::interrupt(),
+        SignalKind::hangup(),
+        SignalKind::terminate(),
+        SignalKind::quit(),
+    ]
+    .into_iter()
+    .map::<anyhow::Result<_>, _>(|kind| {
+        let signal = signal(kind).with_context(|| {
+            format!("Error initializing listener for signal `{kind:?}`")
+        })?;
+        Ok((signal, kind))
+    })
+    .try_collect()?;
+    let futures = signals
+        .into_iter()
+        .map(|(mut signal, kind)| async move {
+            signal.recv().await;
+            info!(?kind, "Received signal");
+        })
+        .map(FutureExt::boxed);
+    future::select_all(futures).await;
+    Ok(())
+}
+
+/// Listen for any exit signals, and return `Ok(())` when any signal is
+/// received. This can only fail during initialization.
+#[cfg(windows)]
+pub async fn signals() -> anyhow::Result<()> {
+    use tokio::signal::windows::{ctrl_break, ctrl_c, ctrl_close};
+
+    let (mut s1, mut s2, mut s3) = (ctrl_c()?, ctrl_break()?, ctrl_close()?);
+    let futures = vec![s1.recv().boxed(), s2.recv().boxed(), s3.recv().boxed()];
+    future::select_all(futures).await;
+    info!("Received exit signal");
+    Ok(())
+}
+
+/// Save some data to disk. This will:
+/// - Ask the user for a path
+/// - Attempt to save a *new* file
+/// - If the file already exists, ask for confirmation
+/// - If confirmed, overwrite existing
+pub async fn save_file(
+    messages_tx: MessageSender,
+    default_path: Option<String>,
+    data: Vec<u8>,
+) -> anyhow::Result<()> {
+    // If the user closed the prompt, just exit
+    let Some(path) =
+        prompt(&messages_tx, "Enter a path for the file", default_path).await
+    else {
+        return Ok(());
+    };
+
+    // If the user input nothing, assume they just want to exit
+    if path.is_empty() {
+        return Ok(());
+    }
+
+    let result = {
+        // Attempt to open the file *if it doesn't exist already*
+        let result = OpenOptions::new()
+            .create_new(true)
+            .write(true)
+            .open(&path)
+            .await;
+
+        match result {
+            Ok(file) => Ok(file),
+            // If the file already exists, ask for confirmation to overwrite
+            Err(error) if error.kind() == io::ErrorKind::AlreadyExists => {
+                warn!(path, "File already exists, asking to overwrite");
+
+                // Hi, sorry, follow up question. Are you sure?
+                if confirm(
+                    &messages_tx,
+                    format!("`{path}` already exists, overwrite?"),
+                )
+                .await
+                {
+                    // REALLY attempt to open the file
+                    OpenOptions::new()
+                        .create(true)
+                        .write(true)
+                        .open(&path)
+                        .await
+                } else {
+                    return Ok(());
+                }
+            }
+            Err(error) => Err(error),
+        }
+    };
+
+    debug!(?path, bytes = data.len(), "Writing to file");
+    let mut file = result
+        .with_context(|| format!("Error opening file `{path}`"))
+        .traced()?;
+    file.write_all(&data)
+        .await
+        .with_context(|| format!("Error writing to file `{path}`"))
+        .traced()?;
+
+    // It might be nice to show the full path here, but it's not trivial to get
+    // that. The stdlib has fs::canonicalize, but it does more than we need
+    // (specifically it resolves symlinks), which might be confusing
+    messages_tx.send(Message::Notify(format!("Saved to {path}")));
+    Ok(())
+}
+
+/// Ask the user for some text input and wait for a response. Return `None` if
+/// the prompt is closed with no input.
+async fn prompt(
+    messages_tx: &MessageSender,
+    message: impl ToString,
+    default: Option<String>,
+) -> Option<String> {
+    let (tx, rx) = oneshot::channel();
+    messages_tx.send(Message::PromptStart(Prompt {
+        message: message.to_string(),
+        default,
+        sensitive: false,
+        channel: tx.into(),
+    }));
+    // Error indicates no response, we can throw that away
+    rx.await.ok()
+}
+
+/// Ask the user a yes/no question and wait for a response
+async fn confirm(messages_tx: &MessageSender, message: impl ToString) -> bool {
+    let (tx, rx) = oneshot::channel();
+    let confirm = Confirm {
+        message: message.to_string(),
+        channel: tx.into(),
+    };
+    messages_tx.send(Message::ConfirmStart(confirm));
+    // Error means we got ghosted :( RUDE!
+    rx.await.unwrap_or_default()
+}